added the changes as mentioned in the comments
