+++ src/utils.cpp
@@ -6,6 +6,10 @@
 #include <tuple>
 #include <vector>
 
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
+namespace py = pybind11;
+
 using AnchorMat = std::vector<std::vector<float>>;
 using Anchor = std::vector<float>;
 
@@ -16,32 +20,6 @@ float softmax(std::vector<float> row) {
   return row[0] / exp_row_sum;
 }
 
-template <typename T>
-std::vector<T> get_items_from_indices(std::vector<T> data,
-                                      std::vector<int> indices) {
-  std::vector<T> output;
-
-  for (auto i : indices) {
-    output.emplace_back(data[i]);
-  }
-  return output;
-}
-// AnchorMat convert_pylist_to_arr(py::list proposals) {
-//   unsigned list_len = py::len(proposals);
-
-//   AnchorMat proposal_arr;
-//   for (int i = 0; i < list_len; ++i) {
-//     Anchor temp;
-//     int col = i % proposal_dim;
-//     temp[col] = proposals[i].cast<float>();
-//     if (col == proposal_dim - 1) {
-//       proposal_arr.emplace_back(temp);
-//     }
-//   }
-
-//   return proposal_arr;
-// }
-
 std::vector<int> create_sequence_vector(int start, int end) {
   // could use directly in the function. It is one line
   //
@@ -50,19 +28,14 @@ std::vector<int> create_sequence_vector(int start, int end) {
   return result;
 }
 
-template <int Cols>
-AnchorMat extract_columns(const AnchorMat &arr,
-                          const std::vector<int> &columns) {
-  AnchorMat extracted_columns(arr.size(), std::vector<float>(columns.size()));
-
-  for (size_t i = 0; i < arr.size(); ++i) {
-    for (size_t j = 0; j < columns.size(); ++j) {
-      extracted_columns[i][j] = arr[i][columns[j]];
-    }
-  }
-  return extracted_columns;
-}
-
+// template std::vector<int>
+// get_items_from_indices(const std::vector<int> &data,
+//                        const std::vector<int> &indices);
+// template std::vector<float>
+// get_items_from_indices(const std::vector<float> &data,
+//                        const std::vector<int> &indices);
+// template AnchorMat extract_columns<20>(const AnchorMat &arr,
+//                                        const std::vector<int> &columns);
 std::vector<std::vector<float>>
 FilterRows(const std::vector<std::vector<float>> &data,
            const std::vector<int> &order) {
@@ -79,18 +52,6 @@ FilterRows(const std::vector<std::vector<float>> &data,
   return other_data;
 }
 
-template <typename T> std::vector<int> argsort(const std::vector<T> &v) {
-  // Initialize original index locations
-  std::vector<int> idx(v.size());
-  std::iota(idx.begin(), idx.end(), 0);
-
-  // Sort indices based on comparing values in v
-  std::sort(idx.begin(), idx.end(),
-            [&v](int i1, int i2) { return v[i1] > v[i2]; });
-
-  return idx;
-}
-
 void get_new_order(std::vector<int> &order, std::vector<int> indices) {
 
   std::vector<int> new_order;
@@ -102,36 +63,6 @@ void get_new_order(std::vector<int> &order, std::vector<int> indices) {
       continue;
   }
   order = new_order;
-}
-
-std::vector<int> distance_greater_indices(const std::vector<float> &acc_dis,
-                                          const float &threshold,
-                                          const std::vector<int> &order) {
-
-  std::vector<int> order_copy = order;
-  std::sort(order_copy.begin() + 1, order_copy.end());
-  order_copy.erase(order_copy.begin());
-
-  std::vector<int> indices;
-  for (int i = 0; i < acc_dis.size(); ++i) {
-    if (acc_dis[i] > threshold) {
-      indices.emplace_back(order_copy[i]);
-    }
-  }
-  return indices;
-}
-
-std::vector<float> compute_score(const AnchorMat &proposals) {
-  std::vector<float> scores;
-
-  std::vector<int> indx = create_sequence_vector(65, 85);
-  std::vector<std::vector<float>> scores_before_softmax =
-      extract_columns<proposal_dim>(proposals, indx);
-
-  for (auto row : scores_before_softmax) {
-    scores.emplace_back(1 - softmax(row));
-  }
-  return scores;
 }
 
 std::vector<float> distance_between_2_lanes(std::vector<float> x1,
@@ -150,25 +81,6 @@ std::vector<float> distance_between_2_lanes(std::vector<float> x1,
   return distance;
 }
 
-std::vector<float>
-compute_all_lane_distance(const std::vector<std::vector<float>> &all_x,
-                          const std::vector<std::vector<float>> &all_z,
-                          const std::vector<float> &x1,
-                          const std::vector<float> &z1) {
-  std::vector<float> distance_arr;
-
-  for (int i = 0; i < all_x.size(); ++i) {
-    std::vector<float> x2 = all_x[i];
-    std::vector<float> z2 = all_z[i];
-    std::vector<float> dis = distance_between_2_lanes(x1, x2, z1, z2);
-    float accumulated_dis = std::accumulate(dis.begin(), dis.end(), 0.0);
-
-    accumulated_dis /= (float)predication_steps;
-    distance_arr.emplace_back(accumulated_dis);
-  }
-  return distance_arr;
-}
-
 std::vector<float> test_scores(const AnchorMat &proposals_arr) {
   std::vector<float> scores;
   for (auto i : proposals_arr) {
@@ -177,29 +89,17 @@ std::vector<float> test_scores(const AnchorMat &proposals_arr) {
   return scores;
 }
 
-std::tuple<AnchorMat, std::vector<float>, std::vector<int>>
-filter_proposals(const AnchorMat &proposals_arr, float conf_threshold,
-                 bool is_test) {
-  std::vector<int> anchor_inds = create_sequence_vector(0, num_anchors - 1);
-  AnchorMat proposals_after_thresholding;
-  std::vector<float> scores_after_thresholding;
-  std::vector<int> anchor_inds_after_thresholding;
-  std::vector<float> scores;
-  if (is_test) {
-    scores = test_scores(proposals_arr);
-  } else {
-    scores = compute_score(proposals_arr);
+void PrintArr(std::array<int, 4431> arr) {
+  for (auto i : arr) {
+    std::cout << i << std::endl;
   }
+}
 
-  for (size_t i = 0; i < scores.size(); ++i) {
-    if (scores[i] > conf_threshold) {
-      proposals_after_thresholding.emplace_back(proposals_arr[i]);
-      scores_after_thresholding.emplace_back(scores[i]);
-      anchor_inds_after_thresholding.emplace_back(anchor_inds[i]);
+void PrintMat(const AnchorMat &mat) {
+  for (auto i : mat) {
+    for (auto j : i) {
+      std::cout << std::setw(3) << j << "  ";
     }
+    std::cout << std::endl;
   }
-
-  return std::make_tuple(proposals_after_thresholding,
-                         scores_after_thresholding,
-                         anchor_inds_after_thresholding);
 }
