+++ tests/CMakeLists.txt
@@ -4,7 +4,9 @@ set(TEST_BINARY ${PROJECT_NAME}_test)
 
 add_executable(${TEST_BINARY}
                test_nms.cpp
+               utils.cpp
                )
+include_directories(../include)
 
 target_link_libraries(${TEST_BINARY} ${PROJECT_NAME} GTest::Main)
 gtest_discover_tests(${TEST_BINARY} WORKING_DIRECTORY
